Age grade: change update source
When you last updated the percent to a specific value, you do not want
changing the gender to change the percent. Instead, you want the time to
update accordingly. As such, keep track of which of the two was last
changed and update the other if any of the age/gender/distance factors
change.